Create and delete Todo per user

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma';
 import { NextResponse, NextRequest } from 'next/server'
 import * as yup from 'yup';
@@ -31,10 +32,15 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
 
+  const user = await getUserSessionServer();
+  if( !user ){
+    return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
+  }
+
   try {
     const { description, completed } = await postSchema.validate ( await request.json() );
 
-    const todo = await prisma.todo.create({ data: { description, completed } });
+    const todo = await prisma.todo.create({ data: { description, completed, userId: user.id} });
   
     return NextResponse.json(todo);
 
@@ -45,9 +51,14 @@ export async function POST(request: Request) {
 
 export async function DELETE(request: Request) {
 
+  const user = await getUserSessionServer();
+  if( !user ){
+    return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
+  }
+
   try {
 
-    await prisma.todo.deleteMany({ where: {completed: true} });
+    await prisma.todo.deleteMany({ where: {completed: true, userId: user.id } });
   
     return NextResponse.json("Deleted all completed todos");
